Bug 916845: Use size_t to describe length of source code in SpiderMonkey SourceHook lazy source hook. r=jorendorff

+++ js/src/jsfriendapi.h
@@ -204,7 +204,7 @@ extern JS_FRIEND_API(bool)
 JS_DefineFunctionsWithHelp(JSContext *cx, JSObject *obj, const JSFunctionSpecWithHelp *fs);
 
 typedef bool (* JS_SourceHook)(JSContext *cx, const char *filename,
-                               jschar **src, uint32_t *length);
+                               jschar **src, size_t *length);
 
 extern JS_FRIEND_API(void)
 JS_SetSourceHook(JSRuntime *rt, JS_SourceHook hook);
